Replaced strict ipam by filteripam implementation
